Compute sector counts without integer overflow
Phase 0 replaced a double-based ceiling division with the int form
(size + sectorSize - 1) / sectorSize, which overflows negative for a
file within one sector of Integer.MAX_VALUE: about 2 GiB with 4 KiB
sectors, or just over 1 GiB at the largest sector size the format
allows. The sector offset entry count then comes out invalid and
extraction fails on an archive the old code read fine, so this was a
regression.

Extracted the division into MpqFile.sectorCount, done in long
arithmetic, and used it from all three call sites instead of repeating
it. Same overflow shape fixed in ZlibStore and JzLibHelper's
worst-case sizing.

Being a named method makes the boundaries testable, which they were not
while the expression was inlined three times.

+++ src/main/java/systems/crigges/jmpq3/compression/JzLibHelper.java
@@ -144,7 +144,7 @@ public final class JzLibHelper {
      * block covers at most 65535 bytes and costs 5 bytes of header.
      */
     private static int worstCaseZlibSize(int n, boolean zlibWrapper) {
-        final int blocks = (n + 65534) / 65535;
+        final int blocks = (int) (((long) n + 65534) / 65535);
         final int header = zlibWrapper ? 2 + 4 : 0;
         return n + blocks * 5 + header + 16;
     }
